+++ src/main/kotlin/sidomik/samples/transfermoney/model/Account.kt
@@ -1,12 +1,13 @@
 package sidomik.samples.transfermoney.model
 
 import java.lang.IllegalArgumentException
+import java.math.BigDecimal
 
 data class Account(val name: String,
-                   var balance: Double) {
+                   var balance: BigDecimal) {
     var id: Long = -1
 
-    fun withDraw(amount: Double) {
+    fun withDraw(amount: BigDecimal) {
         if (balance < amount) {
             throw IllegalArgumentException("Account $id has balance=$balance less than amount=$amount")
 
@@ -14,7 +15,7 @@ data class Account(val name: String,
         balance -= amount
     }
 
-    fun deposit(amount: Double) {
+    fun deposit(amount: BigDecimal) {
         balance += amount
     }
 }
