+++ src/main/kotlin/sidomik/samples/transfermoney/model/Transfer.kt
@@ -1,5 +1,7 @@
 package sidomik.samples.transfermoney.model
 
+import java.math.BigDecimal
+
 data class Transfer(val from: Long,
                     val to: Long,
-                    val amount: Double)
+                    val amount: BigDecimal)
